TODO: response dto

+++ src/board/board.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { BoardEntity } from './entity/board.entity';
 import { BoardRepository } from './entity/board.repository';
@@ -16,11 +16,15 @@ export class BoardService {
     @InjectRepository(ReplyEntity) private replyRepository: ReplyRepository,
   ) {}
 
-  getBoard(boardId: number): Promise<BoardDto> {
-    return this.boardRepository.findOne({
+  async getBoard(boardId: number): Promise<BoardDto> {
+    const found = await this.boardRepository.findOne({
       where: { id: boardId },
       relations: ['replys'],
     });
+    if (!found) {
+      throw new NotFoundException(`can't find board data boardId : ${boardId}`);
+    }
+    return found;
   }
 
   getAllBoards(): Promise<BoardDto[]> {
@@ -36,29 +40,74 @@ export class BoardService {
   modifyBoard(boardId: number, boardDto: BoardDto): Promise<UpdateResult> {
     const { writer, title, content } = boardDto;
     const board = BoardEntity.from(writer, title, content);
-    return this.boardRepository.update(boardId, board);
+
+    try {
+      const result = this.boardRepository.update(boardId, board);
+      return result;
+    } catch (e) {
+      return e;
+    }
   }
 
-  deleteBoard(boardId: number): Promise<DeleteResult> {
-    return this.boardRepository.delete({ id: boardId });
+  async deleteBoard(boardId: number): Promise<DeleteResult> {
+    const find = await this.boardRepository.findOne({ where: { id: boardId } });
+    if (!find) {
+      throw new NotFoundException(`can't find board data boardId : ${boardId}`);
+    }
+
+    try {
+      const result = this.boardRepository.delete({ id: boardId });
+      return result;
+    } catch (e) {
+      return e;
+    }
   }
 
   async saveReply(boardId: number, replyDto: ReplyDto): Promise<ReplyDto> {
     const { writer, content } = replyDto;
+
     const board = await this.boardRepository.findOne({
       where: { id: boardId },
     });
+
+    if (!board) {
+      throw new NotFoundException(`can't find board data boardId : ${boardId}`);
+    }
+
     const reply = ReplyEntity.from(writer, content, board);
-    return this.replyRepository.save(reply);
+    try {
+      const result = this.replyRepository.save(reply);
+      return result;
+    } catch (e) {
+      return e;
+    }
   }
 
   modifyReply(replyId: number, replyDto: ReplyDto): Promise<UpdateResult> {
     const { writer, content } = replyDto;
     const reply = ReplyEntity.from(writer, content);
-    return this.replyRepository.update(replyId, reply);
+    try {
+      const result = this.replyRepository.update(replyId, reply);
+      return result;
+    } catch (e) {
+      return e;
+    }
   }
 
-  deleteReply(replyId: number): Promise<DeleteResult> {
-    return this.replyRepository.delete({ id: replyId });
+  async deleteReply(replyId: number): Promise<DeleteResult> {
+    const reply = await this.replyRepository.findOne({
+      where: { id: replyId },
+    });
+
+    if (!reply) {
+      throw new NotFoundException(`can't find reply data replyId : ${reply}`);
+    }
+
+    try {
+      const result = this.replyRepository.delete({ id: replyId });
+      return result;
+    } catch (e) {
+      return e;
+    }
   }
 }
